fix(downloadCount): fix eslint error
fix eslint error
